tests/pyproject.toml
[project]
name = "interview-agent-tests"
version = "0.1.0"
description = "E2E tests for interview orchestrator"
requires-python = ">=3.11"
dependencies = [
    # Core testing
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-timeout>=2.2.0",
    "pytest-env>=1.1.0",

    # Agent framework (required by orchestrator)
    "google-adk==1.16.0",
    "a2a-sdk>=0.3.0",
    "ap2 @ git+https://github.com/google-agentic-commerce/AP2.git",

    # Database (required by ADK)
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "asyncpg>=0.29.0",

    # HTTP client
    "httpx>=0.28.1",

    # WebSocket client (for orchestrator E2E tests)
    "websockets>=12.0",

    # Server (for starting Google agent)
    "uvicorn>=0.30.0",

    # Environment
    "python-dotenv>=1.0.0",

    # Code quality
    "ruff>=0.8.0",
]

[tool.pytest.ini_options]
# Test discovery
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"

# Async support
asyncio_mode = "auto"

# Logging
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(asctime)s [%(levelname)8s] %(name)s - %(message)s"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"

# Test output
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
]

# Markers
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "e2e: marks tests as end-to-end tests",
]

# Timeout
timeout = 300
